Rework cli module
